html/semantics/selectors/case-sensitivity-of-selectors/selector-namespace-prefix.html

<!doctype html>
<title>Test whether namespace prefixes are lowercased</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
@namespace xhtml "http://www.w3.org/1999/xhtml";

xhtml|span {
    color: rgb(0, 128, 128);
}

XHTML|span {
    color: rgb(255, 165, 0);
}
</style>
<span></span>
<div></div>
<script>
test(function() {
  let span = document.querySelector("span");
  let cs = window.getComputedStyle(span);
  assert_equals(cs.color, "rgb(0, 128, 128)");
}, "element selector");
</script>
